Feacture/user logging system 1 (#12)
* Update: Terminal Session Step 1

Now we reached the starting point to use coroutines as it should(more or less), the next step is to make the TerminalService NoneBlocking, to allow all services to run their steps.

* Just a bunch of changes

* Fixed Logging in

* Changing Task Manager

Moving messages to debugger instead of console and adding an extra try catch block to handle the afterstart of services

* Update ListCommand.cs

List Command now list files and directories on the current directory of the user session.

* Create SystemSession.cs

Added System Session, a user Session that will handle some miscellaneous task in the future, for now it is the one who start the first services to run on the system.

* Moved TerminalSession to it's own file

Also added the logic to load the current directory (default to HOME). and cleaned up the Update method.

* Removed Sha implementations

Since we now use the 'acryptohashnet' nugget package instead.

* Update User.cs

Added a property that gets the UserEntry of the 'users.sys' file.

* Removed embedded resources of no use.

* Moved System.Private.CoreLib.dll to isoFiles

Also moved the file that will be the template for users env file.

* Update WorkFactorCalculatorService.cs

Now Messages are send to the Debugger instead of the Console.

* Update Session.cs

made session public

* Update SessionState.cs

Made public too

* Update Kernel.cs

Since the System Session will start the services, removed the WorkFactorCalculatorService register and changed it to the Session manager.

* Update Kernel.cs

Made Session property public, preparing to make 'Environment' class plugs (The intention is to move out most of the command that do not need to access to Kernel objects direcly and use classes like 'Environment' and so instead).

* Create Zaphyros.conf

Added the template for the user configuration file (this file can and will probably change)

* Update users.sys

Fix typo

* Update pass.conf

Changes to hastype values

* Update UserLogginService.cs

Changes to comply with the new Service launch model (more changes comming on that department)

* Update BCryptSeedProvider.cs

Just a method rename

* Update PasswordConstants.cs

Performing changes made on the pass.conf file

* Update PasswordCheckerService.cs

Messages are now send to the debugger instead of the console.

* Update SessionManagerService.cs

Added static instance, changed message send to the debugger and added missing overrides.